refactor: separate Validator into InputValidator and Validator
